TICKET #206 BUGFIX: fixed the orientation of the spawnpoint and fixed the record JoinGameDTO (it took two ints, but the backend expects two floats and frontend sends floats) to center the newly joined player on a tile

+++ spring/src/main/java/de/hsrm/mi/swt_project/demo/messaging/JoinGameDTO.java
@@ -4,6 +4,6 @@ package de.hsrm.mi.swt_project.demo.messaging;
  * @author Marieke Schmitz 
  * Data Transfer Object that the server receives from the client to request validation of sent gamename
  */
-public record JoinGameDTO(String user, String type, int xPos, int yPos) {
+public record JoinGameDTO(String user, String type, float xPos, float yPos) {
     
 }
